backend/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# CORS handling
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI and Embeddings
openai==1.3.7
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# Machine Learning and NLP dependencies
torch==2.1.1
torchvision==0.16.1
torchaudio==0.16.1
transformers==4.35.2
numpy==1.24.3
scipy==1.11.4

# HTTP and Networking
requests==2.31.0
urllib3==2.0.7

# Data Processing
pandas==2.1.4

# Utilities
pillow==10.1.0
tqdm==4.66.1
click==8.1.7

